obit: Swap plplot dependency for pgplot
The standard version of plplot on Debian / Ubuntu / etc is the
double-precision one resulting in a library named libplplotd instead of
libplplot. This is also the version installed by this tap.

Unfortunately Obit requires the single-precision version, and these
versions cannot co-exist. We therefore switch to the more convenient
pgplot library (Obit can use either of these two libraries) which works
in its standard incarnation. Remove the patch to plplot.m4 as well,
which means that libplplotd will not be picked up even if installed.