New formula: tigger
This is the standalone FITS viewer from the MeqTrees suite. Do a manual
install of the package part of Tigger to the Python site-packages
directory, while bin and docs go to the usual locations and the icons
go to a common MeqTrees share directory. The two sets of patches get
rid of an X11-specific method call and make all Tigger imports
absolute.