New formula: cattery
This package is part of the MeqTrees suite and provides a bunch of
high-level Python modules. It is quite tightly coupled to Timba aka
meqtrees but given a separate formula as it resides in a separate
part of the MeqTrees repository with its own release cycle. Since
the installation part is trivial this then becomes a dependency for
the meqtrees package.