New formula: owlcat
This package is part of the MeqTrees software suite and provides
high-level Python scripts accessed via a shell wrapper. While it
depends on the meqtrees package in reality, the dependency has been
turned around to allow meqtrees to be the single point of entry for
the suite and also to be able to use this package for testing the
whole suite. The previous release is unsupported due to the presence
of filenames that only differ in case and therefore upsets Mac OS X.