Make cmake build type more explicit
Explicitly set CMAKE_BUILD_TYPE in case we want to change it in
future, e.g. to add debug symbols. Bear in mind that the superenv
enforces a strict release build, so any other build needs the std env.
This is a straight refactor and should not change any installation.