meqtrees: Enable debug build
Since MeqTrees is still quite experimental it is useful to have debug
symbols included for tracing problems. The simplest way is to set the
cmake build type explicitly, replacing the default build type of None.
This also requires the std env as the superenv insists on stripping all
symbols. Use dsymutil to assemble symbol files for all libraries for
starters. This was inspired by mxcl/homebrew commit 00cf15ff23fd and
PR #12384.