pyrap: Support brew Python and others
The SConstruct script in libpyrap pulls in the system Python framework
on Darwin by default but this is undesirable if another Python is in
use. This causes pyrap to build fine but throw cryptic exceptions when
opening a CASA Table ("Expecting an object of type dict; got an object
of type dict instead" - see pyrap issue #20). This build behaviour is
commented out as a workaround but ideally should be redone using
python-config to determine the correct linker flags.