Update formulas to use latest Python incantations
These formulas now use the python object in the Formula class to obtain
the local and global site_packages paths and the Python executable.
This removes the need for the which_python utility function and also
improves support for various versions of System / Homebrew Python
(see Homebrew PR 18588 for more details of the python object).