xpra: Add HEAD version
The HEAD version contains initial support for the OS X clipboard.
Depend on PyObjC for AppKit and optionally add OpenGL support.
The patches remove references to gtkosx_application, a module that
is provided by the gtk-mac-integration library which is currently
problematic for brew (see PR #17490). It also finds a substitute
for the system bell and fixes the icon path.