casarest: Disable clang and patches for now
The latest clang still does not like the GCC extensions in use (e.g.
variable-length arrays). This will need some actual code fixes as
everyone should use clang on Homebrew.

Hard-code libstdc++ to avoid compiler clashes on clang. Ignore the
patches for now as they seem to be obsolete on the stable version
as well (?).