psrchive 2012-12 (new formula)
PSRCHIVE is an Open Source C++ development library for the analysis of
pulsar astronomical data. It implements an extensive range of algorithms
for use in pulsar timing, scintillation studies, polarimetric calibration,
single-pulse work, RFI mitigation, etc.

This version is currently head-only, while some more patching is required
to get the stable but old version 2012-12 to work.