casacore 2.0.1
This now pulls from the new GitHub home of casacore. The hacky download
strategy is not needed anymore, as is the devel block for trying out
the 2.0 branch.

There is no requirement to use C++11, as the Complex type has been fixed.
Python support is included by default, though, as pyrap and python-casacore
won't work otherwise. Both these are turned into options.

Fix the tests to be more brewy. Pass the strict audit.