xpra: Restore dependence on gtk-mac-integration
Now that gtk+ is based on Quartz and not X11 (the monster has been
slayed, see legacy PR #39868) it is possible to use gtkosx_application
again. Remove the relevant patches that disabled it.

Fix the resultant icon directory detection which now tried to use
osx_application and got stuck inside the Python executable directories.
Ignore osx_application for this purpose and fall back to the hardcoded
prefix dir inside the Cellar.