casacore-data: switch from ATNF to Astron version
The latest ATNF measures_data tarball has a timestamp of 30 May 2018
while the Astron WSRT_Measures tarball is up to date.

The only funny thing is the ".ztar" file extension (why not ".tar.gz"?),
which would not unpack via the usual mechanisms in Homebrew. Add a
download strategy for it (based on NoUnzipCurlDownloadStrategy). An
alternative solution is to add ".ztar" to the list of extensions in
Homebrew/Library/Homebrew/unpack_strategy/tar.rb but I'm not sure
anyone else ever used that extension...